fix(frontend): vite/client can not be resolved
vite/client adds types for import.meta which are injected by vite for building/bundling see. But vite suggests to add this type explicitly in the tsconfig.json, which requires to explicitly list all types that shall be used in our frontend, which is annoying and not transparent in my opinion.

The current way of importing it as an es6 module where it is needed didn't work either. So we worked around it and avoided the usage of the import utilities.